pyproject.toml
[project]
name = "g2b"
version = "0.1.0"
description = "Woorisoft G2B tender management system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "sqlmodel>=0.0.24",
  "sqlalchemy>=2.0.0",
  "pydantic-settings>=2.10.0",
  "python-dotenv>=1.1.0",
  "httpx>=0.28.0",
  "jinja2>=3.1.6",
  "playwright>=1.54.0",
]

[dependency-groups]
dev = [
  "pytest>=8.4.0",
  "ruff>=0.11.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
